backend/requirements.txt

gunicorn==20.1.0
MarkupSafe==2.1.0
Jinja2==3.0.3
itsdangerous==2.1.0
Werkzeug==2.0.3
Flask==2.0.3
python-dotenv
line-bot-sdk==2.0.1
pymongo[srv]==3.12.0
Flask-APScheduler==1.12.3
Flask-WTF==0.15.1
WTForms==2.3.3
pandas==1.2.4
openpyxl
gevent==21.12.0

